requirements.txt
anthropic>=0.52.0
pdfplumber==0.11.4
python-docx==1.1.2
python-dotenv==1.0.1
